Rename index.js --> speed-convert.js; Add index.js
